Refactor ThankYouHtml into helper
